pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cmdmesh"
version = "0.1.0"
description = "CLI tool to make your life better"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12.3",
    "rich>=13.7.1",
    "httpx>=0.27.0",
    "cryptography>=42.0.8",
    "keyring>=25.2.1",
    "pydantic>=2.7.4",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "sqlalchemy>=2.0.31",
    "asyncpg>=0.29.0",
    "alembic>=1.13.2",
    "redis>=5.0.7",
    "bcrypt>=4.1.3",
    "PyJWT>=2.8.0",
    "aiosmtplib>=3.0.1",
    "pydantic-settings>=2.3.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.2.0",
    "celery[redis]>=5.4.0",
    "kombu>=5.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "fakeredis>=2.23.3",
    "aiosqlite>=0.20.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.5.1",
    "ruff>=0.4.0",
]

[project.scripts]
cmdmesh = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["cli", "backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=backend --cov=cli --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", "backend/db/migrations/*"]
